Move ItemByIdDataloader as we don't need a dedicated folder
