+++ GraphQL/Item/ItemQueries.cs
@@ -2,7 +2,6 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using HotChocolate.Types;
-using WeDoTakeawayAPI.GraphQL.Item.DataLoaders;
 
 namespace WeDoTakeawayAPI.GraphQL.Item
 {
